requirements.txt
Django==4.2.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
pgvector
drf-spectacular
redis==5.0.1
celery==5.3.4
openai==1.12.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
